refactor: refactor common native build config parsing logic with prepareNativeBuild.ts
